+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## [0.5.0] - 2024-10-27
+
 ### Added
 
 * A new theme variant `Pastel Evening Dark #2` which is a mix of orange/yellow shades with a bit of blue for functions.
@@ -175,7 +177,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - Initial release
 
-[Unreleased]: https://github.com/kaiusl/pastel_evening_vscode/compare/v0.4.3...HEAD
+[Unreleased]: https://github.com/kaiusl/pastel_evening_vscode/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/kaiusl/pastel_evening_vscode/releases/tag/v0.5.0
 [0.4.3]: https://github.com/kaiusl/pastel_evening_vscode/releases/tag/v0.4.3
 [0.4.2]: https://github.com/kaiusl/pastel_evening_vscode/releases/tag/v0.4.2
 [0.4.1]: https://github.com/kaiusl/pastel_evening_vscode/releases/tag/v0.4.1